Make projects wrap on smaller displays
The project list fundamentally is a 1-dimensional
list of projects that should wrap when the space
runs out. In cases like this, flexbox is actually
easier to use than grid.
As before, when there is enough space, 4 projects
are displayed next to each other but now on smaller
screens, the number of projects per row shrinks
so that they can still fit the container without
overflowing.