Use `margin` instead of `gap` for flexbox in Safari
Safari does not yet support the `gap` property in flexbox
contexts which made the project list look very crammed.
We can achieve the same behaviour by using a regular
margin though. Since there will now be two margins next to
each other (left, right) insted of just one gap-made "margin",
we need to set the margin to half of what we had set it before
when it was a `gap`. This leads to the same visual display
in all modern browsers.