Improve project list display on low resolutions (#185)
* Make projects wrap on smaller displays

The project list fundamentally is a 1-dimensional
list of projects that should wrap when the space
runs out. In cases like this, flexbox is actually
easier to use than grid.
As before, when there is enough space, 4 projects
are displayed next to each other but now on smaller
screens, the number of projects per row shrinks
so that they can still fit the container without
overflowing.

* Let project images grow to fill project card

* Use `margin` instead of `gap` for flexbox in Safari

Safari does not yet support the `gap` property in flexbox
contexts which made the project list look very crammed.
We can achieve the same behaviour by using a regular
margin though. Since there will now be two margins next to
each other (left, right) insted of just one gap-made "margin",
we need to set the margin to half of what we had set it before
when it was a `gap`. This leads to the same visual display
in all modern browsers.

* Remove unused get-started page and content